fix(transform): preserve complex parameters and add complex64 regression test in scale_by_adamax
- Use `m / denom` directly to preserve imaginary component when `m` is complex.
- Retain float32 promotion for float16 safe division.
- Added `test_adamax_complex_parameters` regression test.

+++ optax/_src/transform.py
@@ -431,7 +431,7 @@ def scale_by_adamax(
       # whenever `m` and `v` are also 0 (e.g. on a zero-gradient step).
       safe_v = v.astype(jnp.promote_types(v.dtype, jnp.float32))
       denom = jnp.maximum(safe_v, eps)
-      return (m.astype(denom.dtype) / denom).astype(m.dtype)
+      return (m / denom).astype(m.dtype)
 
     updates = jax.tree.map(_update, mu_hat, nu, is_leaf=lambda x: x is None)
     return updates, ScaleByAdamState(count=count_inc, mu=mu, nu=nu)
